/stocks/buy implemented & tested.

+++ src/main/java/com/example/assesment/service/UserService.java
@@ -14,6 +14,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import javax.servlet.http.HttpServletRequest;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.regex.Pattern;
@@ -53,7 +54,7 @@ public class UserService {
 
         if (!userRepository.existsByEmail(user.getEmail())) {
             user.setPassword(passwordEncoder.encode(user.getPassword()));
-            user.setBalance(5000.0);
+            user.setBalance(BigDecimal.valueOf(5000.0));
             user.setRoles(new ArrayList<>(Arrays.asList(Role.ROLE_USER)));
             userRepository.save(user);
             return jwtProvider.createToken(user.getEmail(), user.getRoles());
